Version update to 1.0
TiledSharp has not been tagged for a long time, and has since had many
changes both internally and to support later TMX changes.

The versioning scheme has also traditionally tried to match the
compatible version of Tiled.  But that may no longer be practical or
necessary, now that Tiled has had its 1.0 release.

Therefore, I am tagging this as "1.0" as a fresh start, to gather all of
the existing changes since 0.15 and to prepare for future updates.